+++ packages/safe-apps-sdk/src/eth/index.ts
@@ -41,6 +41,8 @@ class Eth {
   public getTransactionReceipt;
   public getTransactionCount;
   public getGasPrice;
+  public getEstimateGas;
+
   private readonly communicator: Communicator;
 
   constructor(communicator: Communicator) {
@@ -88,6 +90,10 @@ class Eth {
     this.getGasPrice = this.buildRequest<never[], string>({
       call: RPC_CALLS.eth_gasPrice,
     });
+    this.getEstimateGas = (transaction: TransactionConfig): Promise<number> =>
+      this.buildRequest<[TransactionConfig], number>({
+        call: RPC_CALLS.eth_estimateGas,
+      })([transaction]);
   }
 
   private buildRequest<P = never[], R = unknown>(args: BuildRequestArgs) {