+++ packages/safe-apps-sdk/src/safe/index.ts
@@ -1,20 +1,31 @@
 import { ethers } from 'ethers';
-import {
-  EIP_1271_INTERFACE,
-  EIP_1271_BYTES_INTERFACE,
-  MAGIC_VALUE_BYTES,
-  MAGIC_VALUE,
-  calculateMessageHash,
-} from './signatures';
+import { EIP_1271_INTERFACE, EIP_1271_BYTES_INTERFACE, MAGIC_VALUE_BYTES, MAGIC_VALUE } from './signatures';
 import { Methods } from '../communication/methods';
 import { RPC_CALLS } from '../eth/constants';
-import { Communicator, SafeInfo, SafeBalances, GetBalanceParams, RPCPayload, TransactionConfig } from '../types';
+import {
+  Communicator,
+  SafeInfo,
+  ChainInfo,
+  SafeBalances,
+  GetBalanceParams,
+  RPCPayload,
+  TransactionConfig,
+} from '../types';
 
 class Safe {
   private readonly communicator: Communicator;
 
   constructor(communicator: Communicator) {
     this.communicator = communicator;
+  }
+
+  async getChainInfo(): Promise<ChainInfo> {
+    const response = await this.communicator.send<Methods.getChainInfo, undefined, ChainInfo>(
+      Methods.getChainInfo,
+      undefined,
+    );
+
+    return response.data;
   }
 
   async getInfo(): Promise<SafeInfo> {
@@ -100,8 +111,12 @@ class Safe {
     }
   }
 
+  calculateMessageHash(message: string): string {
+    return ethers.utils.hashMessage(message);
+  }
+
   async isMessageSigned(message: string, signature = '0x'): Promise<boolean> {
-    const messageHash = calculateMessageHash(message);
+    const messageHash = this.calculateMessageHash(message);
     const messageHashSigned = await this.isMessageHashSigned(messageHash, signature);
 
     return messageHashSigned;