+++ packages/safe-apps-sdk/src/types/sdk.ts
@@ -1,3 +1,9 @@
+import { ChainInfo as _ChainInfo } from '@gnosis.pm/safe-react-gateway-sdk';
+
+export type ChainInfo = Pick<_ChainInfo, 'chainName' | 'chainId' | 'shortName' | 'nativeCurrency'>;
+
+export { NativeCurrency } from '@gnosis.pm/safe-react-gateway-sdk';
+
 export type BaseTransaction = {
   to: string;
   value: string;
@@ -32,4 +38,8 @@ export type SafeInfo = {
   chainId: number;
   threshold: number;
   owners: string[];
+};
+
+export type PostMessageOptions = {
+  transfer?: any[];
 };