+++ packages/safe-apps-provider/CHANGELOG.md
@@ -1,5 +1,55 @@
 # @gnosis.pm/safe-apps-provider
 
+## 0.10.1
+
+### Patch Changes
+
+- Updated dependencies [43651c0]
+  - @gnosis.pm/safe-apps-sdk@7.0.0
+
+## 0.10.0
+
+### Minor Changes
+
+- 45199f6: Use safe-react-gateway for gateway types. Bump dependencies
+
+### Patch Changes
+
+- Updated dependencies [45199f6]
+  - @gnosis.pm/safe-apps-sdk@6.3.0
+
+## 0.9.3
+
+### Patch Changes
+
+- Updated dependencies [d1ef969]
+  - @gnosis.pm/safe-apps-sdk@6.2.0
+
+## 0.9.2
+
+### Patch Changes
+
+- Updated dependencies [0ee4c94]
+  - @gnosis.pm/safe-apps-sdk@6.1.1
+
+## 0.9.1
+
+### Patch Changes
+
+- Updated dependencies [87984be]
+  - @gnosis.pm/safe-apps-sdk@6.1.0
+
+## 0.9.0
+
+### Minor Changes
+
+- 8c1d1e4: Extend the provider from Node.js EventEmitter class
+
+### Patch Changes
+
+- Updated dependencies [630800f]
+  - @gnosis.pm/safe-apps-sdk@6.0.0
+
 ## 0.8.1
 
 ### Patch Changes