+++ packages/safe-apps-web3modal/CHANGELOG.md
@@ -1,5 +1,58 @@
 # @gnosis.pm/safe-apps-web3modal
 
+## 9.0.0
+
+### Patch Changes
+
+- Updated dependencies [43651c0]
+  - @gnosis.pm/safe-apps-sdk@7.0.0
+  - @gnosis.pm/safe-apps-provider@0.10.1
+
+## 8.0.0
+
+### Minor Changes
+
+- 45199f6: Use safe-react-gateway for gateway types. Bump dependencies
+
+### Patch Changes
+
+- Updated dependencies [45199f6]
+  - @gnosis.pm/safe-apps-provider@0.10.0
+  - @gnosis.pm/safe-apps-sdk@6.3.0
+
+## 7.0.0
+
+### Patch Changes
+
+- Updated dependencies [d1ef969]
+  - @gnosis.pm/safe-apps-sdk@6.2.0
+  - @gnosis.pm/safe-apps-provider@0.9.3
+
+## 6.0.1
+
+### Patch Changes
+
+- Updated dependencies [0ee4c94]
+  - @gnosis.pm/safe-apps-sdk@6.1.1
+  - @gnosis.pm/safe-apps-provider@0.9.2
+
+## 6.0.0
+
+### Patch Changes
+
+- Updated dependencies [87984be]
+  - @gnosis.pm/safe-apps-sdk@6.1.0
+  - @gnosis.pm/safe-apps-provider@0.9.1
+
+## 5.0.0
+
+### Patch Changes
+
+- Updated dependencies [8c1d1e4]
+- Updated dependencies [630800f]
+  - @gnosis.pm/safe-apps-provider@0.9.0
+  - @gnosis.pm/safe-apps-sdk@6.0.0
+
 ## 4.0.0
 
 ### Patch Changes